Move sorting note list method to preference utilities class
