+++ Simplenote/src/main/java/com/automattic/simplenote/NoteListFragment.java
@@ -56,7 +56,6 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.simperium.client.Bucket;
 import com.simperium.client.Bucket.ObjectCursor;
 import com.simperium.client.Query;
-import com.simperium.client.Query.SortType;
 
 import java.util.ArrayList;
 import java.util.Calendar;
@@ -423,9 +422,7 @@ public class NoteListFragment extends ListFragment implements AdapterView.OnItem
         }
 
         query.include(Note.PINNED_INDEX_NAME);
-
-        sortNoteQuery(query);
-
+        PrefUtils.sortNoteQuery(query, requireContext());
         return query.execute();
     }
 
@@ -520,31 +517,6 @@ public class NoteListFragment extends ListFragment implements AdapterView.OnItem
 
     public boolean hasSearchQuery() {
         return mSearchString != null && !mSearchString.equals("");
-    }
-
-    public void sortNoteQuery(Query<Note> noteQuery) {
-        noteQuery.order("pinned", SortType.DESCENDING);
-        int sortPref = PrefUtils.getIntPref(getActivity(), PrefUtils.PREF_SORT_ORDER);
-        switch (sortPref) {
-            case 0:
-                noteQuery.order(Note.MODIFIED_INDEX_NAME, SortType.DESCENDING);
-                break;
-            case 1:
-                noteQuery.order(Note.MODIFIED_INDEX_NAME, SortType.ASCENDING);
-                break;
-            case 2:
-                noteQuery.order(Note.CREATED_INDEX_NAME, SortType.DESCENDING);
-                break;
-            case 3:
-                noteQuery.order(Note.CREATED_INDEX_NAME, SortType.ASCENDING);
-                break;
-            case 4:
-                noteQuery.order(Note.CONTENT_PROPERTY, SortType.ASCENDING);
-                break;
-            case 5:
-                noteQuery.order(Note.CONTENT_PROPERTY, SortType.DESCENDING);
-                break;
-        }
     }
 
     /**
